STY: store tL/tR/bL/bR coords in (reusable) arrays
reduces amount of redundant computations and cleans up code.

this highlights that all 3 of the thick-line drawing things (both
rect layout ones, and the unrooted layout one) are all doing the same
thing after they have these coordinates ready. would be worth creating
a util function that adds the "triangles" to a coords array or
something.