BUG/ENH: set 'for' for layout labels correctly
Should be to the corresponding input ID, not name. Huh!

This has the cool side effect of letting you use the label like the
input -- you can hover over the label to shade in the input, and you
can click on the label to "click on" the input (i.e. change layout).