BUG/ENH: Fix discrete colormaps (close #137)
This replaces Colorer.__colorer with Colorer.__colorArray, which
is what it sounds like on the tin (an array of hex color codes) :)

The downside is that this treats quant. color schemes the same
way as discrete color palettes: so these palettes are just looped,
rather than being scaled to fit the entire range of items.

I'm of the opinion that applying quant. color schemes to discrete
data is a bit useless, so I think the next step is adding some sort
of button / checkbox / etc. that lets the user select if they want
color setting to be done with the looping method (added here) or
the scaling method (as was done previously). Will need to think about
what provides the best UX (I think these defaults will be better
than before, at least).