BUG: Fix funky WebGL coloring bug
See test comments for explanation. TLDR: WebGL (or maybe some
intermediate utility) misinterprets the opacity component in the
color array, which breaks things.

ACTUALLY: thinking about it now, I think this bug might actually be
due to Empress' code, which mixes up colors and coordinates into
a single array when drawing stuff. I bet that that's it -- the "1" is
probably interpreted as a coordinate, which is why things look so
weird.