MNT: Move keepUniqueKeys() to util module
I assume this function could be useful for other contexts, and porting
it to this module should make testing it a bit easier (which will be
useful when refactoring it; #147).

I haven't actually added tests for this function yet, mostly because
I'm having a hard time following it / understanding what it's doing.