MNT/TST: Test color scaling; bump chroma to v2.1
Turns out that the Chroma.js version we were using here was ~4 years
out of date -- I noticed small discrepancies (like, on the order of
~0x01 precision for R/G/B channels of the hex colors) between our
Chroma.js and the Chroma.js docs when writing this test.

It looks like Chroma.js version 1.3.3 improved precision, which is
probably the (main?) cause of the observed discrepancies.

I confirmed that updating the version fixed the test errors.
This implies that we should bump the Chroma version for Emperor
/ other libraries using old versions of Chroma also.