BUG: prevent sample/ftr opts simultaneously shown
getting that commit title down to 50 characters was a struggle,
but you get the idea.

This involved adding back some functions to side panel handler, but
these are a lot more clear and concise so I don't mind them as much.
or maybe i'm just biased because i wrote them lol