BUG: fix bug i just introduced in lw corner case
If thickening *was* done previously, but wasn't currently being
done, then Empress._currentLineWidth wasn't being updated to 0.
This meant that, when redrawing the tree after updating the layout,
an old line width was being used.

Now things are good! yay. (This is the sort of case we should probs
add a test for. tagging #142 accordingly.)