ENH: compute area & scale LW; level -> lw in code
Not suuper satisifed with how the scaled line widths look, or the
fact that I had to make the coefficient 1 / ten million to get
things to not look janky. wondering if there's a better way to
do this...

but this at least works. Still need to add tests, but once that's
done #144 should be closeable