TST: finish up tree area tests
Still doesn't consider the c0 coordinates. Is that a problem? No, I
don't think so, since for any node with a c1 coordinate of (x1, y1)
and a c0 coordinate of (x0, y0), the c0 coordinate will be "further
in" towards the center of the circular layout at (0, 0) than the c1
coordinate -- but there will also be a parent node with a c1 also
"further in", and the root's c0 = c1 at (0, 0) anyway.

Thinking about this kinda makes my head hurt tbqh