TST/MNT: Update QUnit & mv jQuery code
This updates QUnit from v1.12.0 to v1.23.1. There are later versions
of QUnit, but they're of version 2.x and higher which seems to be
incompatible with the way we're using QUnit (I think the main
difference is that QUnit 1.x just injects stuff like "module", "ok",
etc. into the global namespace, while QUnit 2.x keeps this stuff bound
-- e.g. "QUnit.module", etc.).

It'd be nice to update QUnit to the absolute latest versions available
but for now the latest 1.x version is a decent compromise. It at least
lets us use "notOk" in tests.

Also, jQuery was previously stored with normal support files vendor
code, but wasn't being used there (it's only in use with the tests).
To make this clearer, I've moved the jQuery code to the test vendor
files (alongside the QUnit stuff).

For reference, the jQuery version we're on -- 2.2.4 -- seems to be
the latest version of the 2.x series, so it seems at least kinda up-
to-date as well.