ENH: Add spectrum.js; init colorpickers; cycle qcs
"qcs" being "QIIME Colors" but abbreviated because of the 50-char
limit

yeah so i suspect the spectrum stuff will need to be redone in a bit,
i.e. to prevent css messing with emperor spectrum's ... and to split
up vendor code btwn css and js. but for a prototype, ok