ENH: inital fm/sm barplot layer toggling
Altered <hr /> insertion a bit (now each layer puts them at the
end not at the top -- with an extra being put above the first
layer -- the effect of this is that now there's a line btwn the
final layer's "remove this layer" row and the "add another layer"
row, which makes it clearer that adding is an operation that works
on a level "higher" than individual layers if that makes sense.

also removed the empress-hr css class since it was going unused

also initial work on splitting barplot layer html creation into
multiple functions. that'll be nice.