ENH: Working prototype of sample metadata barplots
Needs testing, code polishing, etc. -- but exciting.

Note that the colors can differ slightly between these barplots
and what's shown when coloring the tree by sample metadata. The
reason for this is that the sample-metadata assignment of colors in
colorBySampleCat() is done *AFTER* filtering non-unique metadata
values in _projectObservations().

Interestingly, this isn't the case for feature metadata coloring:
that's because colorByFeatureMetadata() assigns colors before
calling _projectObservations().

It would ... be good to make things consistent, one way or another.