TST: support+test circ layout for 1-child nodes
Progress on #184. Since it looks like the unrooted layout handles
these properly, all that's left for this issue should be just not
drawing arcs for internal nodes with 1 child. Should be doable by
modifying the js circular layout to give 1-child nodes -1 for their
arc / angle values or something to signal to empress.js to not bother
with arcs.