ENH: Scale barplot lengths by max displacement
Relevant to #276. The reliance on max displacement (which doesn't
seem to have an analogue in the unrooted layout? or maybe we can sort
of treat it the same way as circ layout, idk) means that this might
be too specific a solution to apply to line width thicknesses, but
it should make using barplots a lot less painful