Configurable barplot borders (#386)
* ENH: Add UI of barplot border options;fix CSS #350

there was some extra v-space when barplots were available due to
layout but 'draw barplots' was unchecked; this was because a <p>
was not hidden (although the one button inside it was). This
discrepancy has since been fixed.

* ENH: Add barplot borders -- close #350

* STY: prettify

* DOC: tidy comment

* ENH: shift around and tidy up #350 UI

makes things look less cluttered. had too many <hr />s previously

* TST: unbreak tests (add new barplot ui eles)

* DOC: add add'tl comment abt freebies

* DOC: add notes about making it go fast

* TST: very basic #350 tests

more would be nice but probs ok for now

* STY: prettify test

* DOC: fix documentation within initBorderOptions()

* Update empress/support_files/js/empress.js

Co-authored-by: Yoshiki Vázquez Baeza <yoshiki@ucsd.edu>

* TST: beef up barplot border init test

* TST: test colorpicker init for barplot borders

* TST: document barplot border init test

Co-authored-by: Yoshiki Vázquez Baeza <yoshiki@ucsd.edu>