Scale barplot thicknesses relative to tree size; makes default layout barplot lengths look thicker; fix a small Colorer bug (#442)
* ENH: Scale barplot lengths by max displacement

Relevant to #276. The reliance on max displacement (which doesn't
seem to have an analogue in the unrooted layout? or maybe we can sort
of treat it the same way as circ layout, idk) means that this might
be too specific a solution to apply to line width thicknesses, but
it should make using barplots a lot less painful

* BUG: precomp sm barplot lyr len&report correctly

* Document layout factor stuff

* add note re: 1.1

* BUG: Fix CLR_COL / "freebie" problem

* Remove unneeded comments re: freebie thing

* Test rgb2gl; prettify

* improve rgb2gl docs

* TST/MNT: Improve/test layout supported & freebies

* Do away with CLR_COL_GL, and document stuff

should address @ElDeveloper comment

* BUG: Also apply barplot unit to border lens

WHOOPS

* TST: fix barplot border test