STY: Fix flake8 line lengths for #473 tests
The specific error code was E501. Annoyingly, ignoring it in the
Makefile's invocation of flake8 triggered MORE errors of a different
type for some reason lmao.

ANYWAY this should close #473. There is still the question of if
the space tradeoff is worth it or not -- but at the very least this
functionality is ready.