ENH: Implement ancestor fm retrieval in JS
Closes #473 and closes #422.

Avoids having to store this stuff in the QZV, which will save
a lot of space. This does still involve storing the "expanded"
taxonomy for a given level all at once in memory, though -- I do
not think this will be horrendous, since it is not that much extra
data all things considered.